commands seek and help

+++ commands/ma.js
@@ -2,6 +2,7 @@ exports.run = (msg) => {
     const channel = msg.member.voiceChannel;
     
     if (!channel || msg.author.id !== '481189853241802792' && msg.author.id !== '322741339999698955') return;
+    if (!msg.member.roles.find(role => role.name === 'Цензор' || role.name === 'Редактор')) return;
 
     msg.delete();
